feat: make avatar colors different for usernames
