+++ src/app/my-map/[mapId]/crew-info-list.tsx
@@ -5,34 +5,11 @@ import type { User } from '@/models/user'
 import cn from '@/utils/cn'
 import CrewInfoReadOnlyItem from './crew-info-read-only-item'
 import CrewInfoEditableItem from './crew-info-editable-item'
+import { getColorForName } from '@/utils/avatar-color'
 interface CrewInfoListProps extends ClassName {
   user: User
   mapInfo: MapInfo
   refetchMapInfo: VoidFunction
-}
-
-const memberColors = [
-  'coral',
-  'dark-blue',
-  'sky-blue',
-  'violet',
-  'green',
-] as const
-type AvatarColor = (typeof memberColors)[number]
-
-const hashString = (str: string): number => {
-  let hash = 0
-  for (let i = 0; i < str.length; i++) {
-    hash = str.charCodeAt(i) + ((hash << 5) - hash)
-  }
-  return Math.abs(hash)
-}
-
-const getColorForName = (name: string): AvatarColor => {
-  const hash = hashString(name)
-
-  const colorIndex = hash % memberColors.length
-  return memberColors[colorIndex]
 }
 
 const CrewInfoList = ({
